requirements.txt
setuptools>=65.5.0
wheel>=0.38.0
pip>=23.0.0
pymupdf>=1.22.0
pillow>=10.0.0
opencv-python-headless>=4.7.0
supabase>=1.0.0
openai>=1.0.0
pytest>=7.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.0.0
numpy>=1.23.0,<2.0.0
